fix: moves RefreshToken to auth domain
