+++ src/main/java/com/lp/v2/domains/auth/RefreshTokenJpaRepository.java
@@ -1,4 +1,4 @@
-package com.lp.v2.security;
+package com.lp.v2.domains.auth;
 
 import org.springframework.data.jpa.repository.JpaRepository;
